Smooth on Add and Smooth All
Added a Smooth All button on the function page.  Clicking this button
will attempt to smooth the entire graph from right to left within its
constrained box.

Adding a point to the graph will now cause the graph to try to smooth
the line around the new point.  This is working properly if the new
point is added to the LEFT of the path's current control point.  Adding
a point to the RIGHT of the path's control point will try to smooth the
graph in an undesired manner.  A check needs to be added (and a new
smooth function) to see which side of the control point you're adding a
new line, and then smooth the graph from right to left or left to right
accordingly.